refactor(cli): fold agent-config helpers into _providers.py
Move all of _agents.py (AgentProfile, build_deadline_preamble,
list_agents, load_agent_profile, _parse_profile, _resolve_profile_path,
_validate_bare_name, _find_lionagi_dir) into _providers.py. Every importer
(agent.py, main.py, orchestrate/_orchestration.py, state/provenance.py,
tests/) is rewired to lionagi.cli._providers. _agents.py deleted.
