+++ lionagi/cli/orchestrate/_orchestration.py
@@ -22,9 +22,15 @@ from lionagi.operations.builder import OperationGraphBuilder
 from lionagi.protocols.generic.log import DataLoggerConfig
 from lionagi.state import provenance as _provenance
 
-from .._agents import AgentProfile, list_agents, load_agent_profile
 from .._logging import hint
-from .._providers import build_imodel_from_spec, parse_model_spec, resolve_persisted_effort
+from .._providers import (
+    AgentProfile,
+    build_imodel_from_spec,
+    list_agents,
+    load_agent_profile,
+    parse_model_spec,
+    resolve_persisted_effort,
+)
 from .._runs import (
     RunDir,
     _resolve_project,
