+++ tests/agent/test_profile_artifacts.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import pytest
 
-from lionagi.cli._agents import _parse_profile
+from lionagi.cli._providers import _parse_profile
 
 
 class TestProfileArtifactDefaults:
